# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Fstudio",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Development Environment"
        },
        {
            "url": "https://fstudiovn.id.vn/api/",
            "description": "Staging  Environment"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "summary": "Xác thực người dùng và tạo mã thông báo JWT",
                "operationId": "a5385ffca73f113abef00ea604d1ca58",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email của người dùng",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Mật khẩu của người dùng",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công"
                    },
                    "401": {
                        "description": "Thông tin đăng nhập không hợp lệ"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "Đăng xuất người dùng",
                "operationId": "46f8a14c8083abb69eed72768aa08629",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    },
                    "401": {
                        "description": "Token đã hết hạn hoặc không hợp lệ"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}